Move storage of access tokens to context and reducer
Now the access token can be easily updated or used
from anywhere within the application, and all com-
ponents that use it will know when it has been up-
dated, and re-render with the new values.

So now upon logging in, the nav bar updates to re-
flect this, and so does anything else that relies
upon it (e.g. users will not list anything) if you
are not logged in.

Storing access tokens in memory and not on the br-
owser session storage is more secure according to
some people on the internet. I guess if an attacker
gained access to the js, it would be a lot easier
for them to find a value stored in session strorage
than in a variable in memory.