Fix: blog, comment, user not loading route data
This was due to the access token no longer being
stored in sessionStorage and instead stored in
memory and passed to react components with reducer
and contexts.

Moved routes from separate file to part of main
App component, so state can be passed from result
of useReducer() to the loader functions, which
themselves return functions to be used as loaders,
with the accessToken baked in. When the state cha-
nges, i.e. when the accessToken is changed, the
useMemo will re-create the router with the new
accessTokens.

Sounds a bit heavy to re-create the router every
time an access token changes, but I'm not sure
what the 'proper' way to do this is. In fact, I'm
not even sure what the 'proper' way to store user
login info is with react. But it works.