Fix: not handling !response.ok scenarios properly
There was a check for response?.ok before calling
response.json(). Since we want to access the json
payload in scenarios where the response has returned
a 400, 401 or 403 code, the response?.ok check was
a mistake as it will be false if the status code is
in the 400 or 500 ranges.