Switch from reducer to useRef for access token storage
Re-renders not desired when access token is updated,
just when user logs in or out. Access token will get
updated many times across one session, so makes sense
to split these up.

Using a ref to store the access token has the advantage
of it being a bit easier to update the current value
directly in authFetch instead of returning a new access
token if one is fetched during the function (i.e. if
auth fails initially), for the react component to update
the reducer's state.

Works well except one problem is that the page display
doesn't automatically update once a user logs out.
The app re-renders, but the page stays the same.
Try to figure out why!