Fix: Make setting errors more consistent across the application
Some react state for holding errors would in some
situations be set to an object of type Error, and
other times be set to a string.

This has been changed. In the future, any time we
use state to store errors, ensure that whatever it
is, is always an Error object, not just a string.
This will make the app more consistent and avoid
some annoying bugs where the jsx is expecting an
Error but receives a string instead.