Fix: blog page not updating comments list if user replies or creates new comment
This is a hacky, horrible solution. The BlogPage
uses the useComments hook to get comments. The
useComments hook uses useAuthFetch, which returns
a http response and an error if present. The useA-
uthFetch hook has a useEffect callback that only
runs if the url, accessRef or logged in status ch-
anges.

When a user creates a new comment or replies to an
existing comment, none of the useEffect dependencies
change, so the useEffect does not re-run.

My current rubbish solution is to set a uuid on the
url which is changed when I want to force a new fetch.

I have looked at changing the dependencies of the
useAuthFetch hook somehow to make this work in a
way that makes sense. Removing the useEffect is not
possible as that leads to an infinite loop.

What is the proper way to do this?