Use authFetch in loaders, so if auth fails they will automatically try to get a new access token
This is possible now as the access tokens are stored
in a ref, instead of in a reducer/state. This means
the ref object can be passed to a function outside of
react, like authFetch, and authFetch can update the
access token. Since JS passes objects by reference,
this will change the value of the object in memory,
not just a copy local to the function. This was not
possible when using state, since reducers/state can
only be used in components; not in loaders or exte-
rnal functions.

Reducers/state are just for stuff that should prompt
a re-render. Since the access tokens will change all
the time behind the scenes, they should never have
been in reducers/state in the first place.