Move from useBlogs/Comments/Users hooks to loaders
This has solved basically all the problems I was having,
with fetch post or update being done and then the page
not updating. Now it reloads the fresh data when the
component uses the useNavigate hook.

All in all using hooks for those things was a big
mistake. For some reason, I got it into my head
that the data should be filtered server side before
it gets sent to the client. That meant every character
typed in the query, the client would make another fetch
request. What a stupid way of doing it.

Now, the loader grabs all the data and it is filtered
clientside. Feels much quicker. No idea what I was
thinking with that previous approach..

However, the loader should not get all blogs - limit
to a particular number, e.g. use pagination to navigate
to the rest, etc. If there were 10,000 blogs that would
be a pretty terrible user experience, waiting for that
to load.