Modify routes so that index goes straight to log-in page
Since this is an admin web client, we don't want a
user that is not signed in to have access to nav bar.
What is the point in showing them links they aren't
allowed to access? So log in page is on its own route,
and all the other routes are children of root, which
contains the header (i.e. website name and navbar).