Make PageNav component generic so not tied to using url search params
This is so we can have multiple PageNavs on one page,
for different sections of the page. E.g. on the user
page, you can flip through different pages of comments
on one section, and different pages of blogs on another.

The useSearchParamsPageNumber hook is just to streamline
some of the required set up to use a page url query
parameter.