Fix: logged in status resets when app is refreshed
Log in status and access code were split into two,
since the access code will expire and change often.
The access code changing would not warrant a re-render.
Whereas, if the user goes from logged out to logged in,
we would want the app to respond to that.

However, if the app was refreshed, the log in status
would be initialised to its inital value within react,
which is false, but the authFetch calls would use the
refresh token in the cookies to automatically get new
access codes - so the navBar would show the user as
logged out, but the fetch calls would all still work.

The refresh token is an http only cookie, so the
server now sets a second cookie called login which
is accessible by scripts. The presence of this cookie
tells us that the user has logged in.

A new hook called useLoginState is initialised to
true if a login cookie is present, and false if not.
It returns the login status and a set function to change
it. If the login state is set to false, the cookie is
cleared.