Server now accepts onlyUnpublished query param
This is more readable than using publishedAt=null.
Checkboxes don't have values, or true or false.
They either have a checked attribute or not, and
on the query string they end up being 'on' or just
not there at all. So in the url it would end up
being publishedAt=on, which doesn't make much sense
at a glance.

So now it will say onlyUnpublished=on which makes
a lot more sense and is more understandable, when
looking at the code, and resulting url.