Fix: loaders not ordering results in a consistent way
This was resulting in the same results appearing
on multiple pages, seemingly at random. The solution
was to use id as a tiebreaker for ordering. E.g.
if you sort by a text field and two entries have
exactly the same text, the additional order by id
will ensure the results are ordered the same way
every time.