Create useClientWidth hook so different components can be shown depending on width of screen
The idea is that the logic for choosing which component
to show can be programmed in js, not split out into css.

Previously I have had styled components that existed just
to have a breakpoint, which would then set different
child components to hidden or visible. This way, the
breakpoint and logic can be entirely in js and not
split awkwardly between js and css.